requirements.txt
torchvision>=0.6.1
torch==1.13.0
numpy>=1.16.5
pillow>=6.2.0
flask>=1.1.1
werkzeug>=0.16.0
pandas>=0.25.1
scikit-learn>=0.21.3
tqdm>=4.36.1
streamlit>=0.74.1
matplotlib>=3.1.1
scikit-image>=0.15.0
opencv-python>=4.4.0.44
openpyxl